hierarchical_supervisor_worker/pyproject.toml
[project]
name = "hierarchical-supervisor-worker"
version = "0.1.0"
description = "Hierarchical supervisor-worker meta-agent with openai-agents-sdk"
authors = [{ name = "Terence Liu", email = "llebox@hotmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "openai-agents>=0.2.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.0",
  "redis>=5.0.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "langgraph>=0.2.0",
  "weave>=0.50.0",
]

[project.scripts]
hsw = "cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["hsw*"]
